+++ Block/Topic/ListPost.php
@@ -19,35 +19,36 @@
  * @license     https://www.mageplaza.com/LICENSE.txt
  */
 
-namespace Mageplaza\Blog\Block\Category;
+namespace Mageplaza\Blog\Block\Topic;
 
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\Phrase;
 use Mageplaza\Blog\Helper\Data;
 use Mageplaza\Blog\Model\ResourceModel\Post\Collection;
+use Mageplaza\Blog\Model\TopicFactory;
 
 /**
- * Class Listpost
- * @package Mageplaza\Blog\Block\Category
+ * Class ListPost
+ * @package Mageplaza\Blog\Block\Topic
  */
-class Listpost extends \Mageplaza\Blog\Block\Listpost
+class ListPost extends \Mageplaza\Blog\Block\ListPost
 {
     /**
-     * @var string
+     * @var TopicFactory
      */
-    protected $_category;
+    protected $_topic;
 
     /**
      * Override this function to apply collection for each type
      *
-     * @return Collection|null
+     * @return Collection
      * @throws NoSuchEntityException
      */
     protected function getCollection()
     {
-        if ($category = $this->getBlogObject()) {
-            return $this->helperData->getPostCollection(Data::TYPE_CATEGORY, $category->getId());
+        if ($topic = $this->getBlogObject()) {
+            return $this->helperData->getPostCollection(Data::TYPE_TOPIC, $topic->getId());
         }
 
         return null;
@@ -58,17 +59,18 @@ class Listpost extends \Mageplaza\Blog\Block\Listpost
      */
     protected function getBlogObject()
     {
-        if (!$this->_category) {
+        if (!$this->_topic) {
             $id = $this->getRequest()->getParam('id');
+
             if ($id) {
-                $category = $this->helperData->getObjectByParam($id, null, Data::TYPE_CATEGORY);
-                if ($category && $category->getId()) {
-                    $this->_category = $category;
+                $topic = $this->helperData->getObjectByParam($id, null, Data::TYPE_TOPIC);
+                if ($topic && $topic->getId()) {
+                    $this->_topic = $topic;
                 }
             }
         }
 
-        return $this->_category;
+        return $this->_topic;
     }
 
     /**
@@ -80,12 +82,12 @@ class Listpost extends \Mageplaza\Blog\Block\Listpost
         parent::_prepareLayout();
 
         if ($breadcrumbs = $this->getLayout()->getBlock('breadcrumbs')) {
-            $category     = $this->getBlogObject();
-            $categoryName = preg_replace('/[^A-Za-z0-9\-]/', ' ', $category->getName());
-            if ($category) {
-                $breadcrumbs->addCrumb($category->getUrlKey(), [
-                    'label' => __($categoryName),
-                    'title' => __($categoryName)
+            $topic     = $this->getBlogObject();
+            $topicName = preg_replace('/[^A-Za-z0-9\-]/', ' ', $topic->getName());
+            if ($topic) {
+                $breadcrumbs->addCrumb($topic->getUrlKey(), [
+                    'label' => __($topicName),
+                    'title' => __($topicName)
                 ]);
             }
         }
@@ -100,21 +102,21 @@ class Listpost extends \Mageplaza\Blog\Block\Listpost
     public function getBlogTitle($meta = false)
     {
         $blogTitle = parent::getBlogTitle($meta);
-        $category  = $this->getBlogObject();
-        if (!$category) {
+        $topic     = $this->getBlogObject();
+        if (!$topic) {
             return $blogTitle;
         }
 
         if ($meta) {
-            if ($this->helperData->getMetaTitleByStoreId($category->getMetaTitle())) {
-                $blogTitle[] = $this->helperData->getMetaTitleByStoreId($category->getMetaTitle());
+            if ($this->helperData->getMetaTitleByStoreId($topic->getMetaTitle())) {
+                $blogTitle[] = $this->helperData->getMetaTitleByStoreId($topic->getMetaTitle());
             } else {
-                $blogTitle[] = ucfirst($category->getName());
+                $blogTitle[] = ucfirst($topic->getName());
             }
 
             return $blogTitle;
         }
 
-        return ucfirst($category->getName());
+        return ucfirst($topic->getName());
     }
 }
