+++ view/frontend/layout/mpblog_category_view.xml
@@ -24,7 +24,7 @@
     <update handle="mpblog_design"/>
     <body>
         <referenceContainer name="content">
-            <block class="Mageplaza\Blog\Block\Category\Listpost" name="mp.blog.category.post.list" template="Mageplaza_Blog::post/list.phtml"/>
+            <block class="Mageplaza\Blog\Block\Category\ListPost" name="mp.blog.category.post.list" template="Mageplaza_Blog::post/list.phtml"/>
         </referenceContainer>
     </body>
 </page>
