+++ view/frontend/layout/mpblog_post_index.xml
@@ -24,7 +24,7 @@
     <update handle="mpblog_design"/>
     <body>
         <referenceContainer name="content">
-            <block class="Mageplaza\Blog\Block\Post\Listpost" name="mpblog.post.list" as="mpblog_post" template="Mageplaza_Blog::post/list.phtml"/>
+            <block class="Mageplaza\Blog\Block\Post\ListPost" name="mpblog.post.list" as="mpblog_post" template="Mageplaza_Blog::post/list.phtml"/>
             <block class="Magento\Framework\View\Element\Template" name="mpblog.copy.right" after="-" template="Mageplaza_Blog::html/copyright.phtml"/>
         </referenceContainer>
     </body>
