+++ view/frontend/layout/mpblog_topic_view.xml
@@ -24,7 +24,7 @@
     <update handle="mpblog_design"/>
     <body>
         <referenceContainer name="content">
-            <block class="Mageplaza\Blog\Block\Topic\Listpost" name="mp.blog.topic.post.list" template="Mageplaza_Blog::post/list.phtml"/>
+            <block class="Mageplaza\Blog\Block\Topic\ListPost" name="mp.blog.topic.post.list" template="Mageplaza_Blog::post/list.phtml"/>
         </referenceContainer>
     </body>
 </page>
